% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_tools.R
\name{add_alpha_diversity}
\alias{add_alpha_diversity}
\title{Adds alpha diversity metrics to a \code{phyloseq} object.}
\usage{
add_alpha_diversity(ps)
}
\arguments{
\item{ps}{A \code{phyloseq} object that contains \code{\link[phyloseq]{sample_data}}.
If there is a tree in the \code{\link[phyloseq]{phy_tree}} slot, this function will
also add \href{http://www.sciencedirect.com/science/article/pii/0006320792912013}{Faith's Phylogenetic Diversity (1992)}
as implemented in \code{picante} with the \code{\link[picante]{pd}} function.}
}
\value{
This function returns the input \code{phyloseq} object with
 \code{Observed}, \code{Chao1}, \code{ACE}, \code{Shannon}, \code{Simpson},
 \code{InvSimpson}, \code{Fisher}, and \code{Faiths_PD} columns in the \code{sample_data()}.
}
\description{
This function adds several alpha diversity metrics to the
\code{\link[phyloseq]{sample_data}} of a \code{phyloseq} object.
This function largely a reimplementation of \code{\link[phyloseq]{estimate_richness}}.
}
\examples{
ps <- add_alpha_diversity(ps)
}
