% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_tools.R
\name{add_lowest_rank}
\alias{add_lowest_rank}
\title{Adds \code{lowest_rank} to a phyloseq object}
\usage{
add_lowest_rank(ps)
}
\arguments{
\item{ps}{A \code{phyloseq} object that contains a \code{\link[phyloseq]{tax_table}}}
}
\value{
This function returns the input \code{phyloseq} object with a \code{lowest_rank} column in the \code{tax_table()}.
If "Species" is present in the \code{tax_table()}, then this function returns <Genus_species>.
If "Species" is NA, this function returns the next highest taxonomic level available.
If more than 1 "Species" is present, this function returns <Genus_species1/species2+nspecies>.
}
\description{
Adds a \code{lowest_rank} column to a taxonomy table in a phyloseq object.
}
\examples{
ps <- add_lowest_rank(ps)
}
