% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_tools.R
\name{add_mdi}
\alias{add_mdi}
\title{Adds Microbial Dysbiosis Index to a phyloseq object}
\usage{
add_mdi(ps)
}
\arguments{
\item{ps}{A \code{phyloseq} object that contains \code{\link[phyloseq]{sample_data}}}
}
\value{
This function returns the input \code{phyloseq} object with a \code{MDI} column in the \code{sample_data()}.
}
\description{
Adds a \code{MDI} column to \code{\link[phyloseq]{sample_data}} in a \code{phyloseq} object.
The Microbial Dysbiosis Index first
proposed by \href{http://www.sciencedirect.com/science/article/pii/S1931312814000638}{Gevers \emph{et al.} (2014)}.
}
\examples{
ps <- add_mdi(ps)
}
