% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteome_tools.R
\name{add_nsaf}
\alias{add_nsaf}
\title{Normalize protein abundances by \code{NSAF}}
\usage{
add_nsaf(ps, length)
}
\arguments{
\item{ps}{A \code{phyloseq} object that contains a \code{\link[phyloseq]{tax_table}}.
The \code{tax_table} must contain protein length as specified with the \code{length} parameter.}

\item{length}{A column in the \code{tax_table} containing the protein length.}
}
\value{
This function returns the input \code{phyloseq} object with a the  \code{otu_table()} normalized
by \code{NSAF}.
The Normalized Spectral Abundance Factor was
proposed by \href{http://pubs.acs.org/doi/pdf/10.1021/pr060161n}{Zybailov \emph{et al.} (2006)}.
NSAF for a protein `k` is the number of spectral counts (`SpC`, the total number of MS/MS spectra)
identifying a protein, `k`, divided by the protein’s length (`L`),
divided by the sum of `SpC`/`L` for all `N` proteins in the experiment.
}
\description{
normalizes proteome data by a \code{Normalized Spectral Abundance Factor}
using the protein length as specified by the \code{length} parameter in the
\code{\link[phyloseq]{tax_table}} of a phyloseq object
The input \code{\link[phyloseq]{otu_table}} should contain proteomic count or intensity values from a mass spec.
}
\examples{
ps <- add_nsaf(ps)
}
