% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{alpha.boot}
\alias{alpha.boot}
\title{Bootstrap cronbach's alpha with BCA confidence intervals}
\usage{
alpha.boot(df, n = 1000)
}
\arguments{
\item{df}{matrix or dataframe (to be converted using trial.mat)}

\item{n}{number of bootstrap samples to draw}
}
\value{
dataframe with alpha and confidence intervals
}
\description{
Bootstrap cronbach's alpha with BCA confidence intervals
}

