% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{calc.alpha}
\alias{calc.alpha}
\title{Calculate cronbach's alpha across a range of test lengths.}
\usage{
calc.alpha(M, method = c("standard", "boot")[1])
}
\arguments{
\item{M}{matrix with subjects on rows and items on columns}

\item{method}{method for calculating confidence intervals (standard or boot)}
}
\value{
dataframe with numitems, alpha, and confidence intervals
}
\description{
Calculate cronbach's alpha across a range of test lengths.
}

