% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{calc.alphadiff}
\alias{calc.alphadiff}
\title{Convenience wrapper for cocron.
Tests hypothesis that cronbachs alpha for one test length versus another are
is equal using Feldt's test (corrected for dependence)}
\usage{
calc.alphadiff(M1, M2, to_df = TRUE)
}
\arguments{
\item{M1}{first matrix to calculate alpha for}

\item{M2}{second matrix}

\item{to_df}{parse cocron output into dataframe}
}
\value{
cocron output or the same output parsed into dataframe
}
\description{
Convenience wrapper for cocron.
Tests hypothesis that cronbachs alpha for one test length versus another are
is equal using Feldt's test (corrected for dependence)
}

