% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{df.from.cocron}
\alias{df.from.cocron}
\title{Convert cocron output (from cocron package) to data.frame}
\usage{
df.from.cocron(out, onerow = FALSE)
}
\arguments{
\item{out}{cocron output}

\item{onerow}{not used}
}
\value{
dataframe with sample size, items, alpha, confidence intervals, and p-value
}
\description{
Convert cocron output (from cocron package) to data.frame
}

