% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteome_tools.R
\name{merge_protein_replicates}
\alias{merge_protein_replicates}
\title{Merge Protein Replicates}
\usage{
merge_protein_replicates(ps, repID, minAbund = 10, maxZeros = 0.5)
}
\arguments{
\item{ps}{A \code{phyloseq} object that contains a variable in \code{\link[phyloseq]{sample_data}}
denoting which samples are replicates of one another.}

\item{repID}{A column in the \code{sample_data} that identifies which samples are replicates of each other and should be merged.
All replicates should have the same value in this column.}

\item{minAbund}{(optional) A column in the \code{sample_data} specifying which samples are replicates to be merged.
This is the minimum value a protein in a set of replicates has to have for it to be considered "real"
Default value = 10.}

\item{maxZeros}{A column in the \code{sample_data} specifying which samples are replicates to be merged.
Default value = 0.5}
}
\value{
This function returns the input \code{phyloseq} object with the sample replicates merged.
}
\description{
This function merges replicate samples from proteomic data
}
\examples{
ps <- merge_protein_replicates(ps, repID="Subject_ID", minAbund=10, maxZeros = 0.5)
}
