% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_tools.R
\name{plot_ordered_bar}
\alias{plot_ordered_bar}
\title{bar plot ordered across samples}
\usage{
plot_ordered_bar(ps, x = "Sample_ID", y = "Abundance", fill = NULL,
  leg_size = 0.5, title = NULL)
}
\arguments{
\item{ps}{A \code{phyloseq} object.}

\item{x}{the varaiable in the \code{\link[phyloseq]{sample_data}} that you want on the x-axis.
Defaults to "Sample_ID".}

\item{y}{the varaiable in the \code{\link[phyloseq]{sample_data}} that you want on the y-axis.
Defaults to "Abundance".}

\item{fill}{the varaiable in the \code{\link[phyloseq]{sample_data}} that you want to color by.}

\item{leg_size}{a number indicating the legend.key.size
The default value is 0.5.}

\item{title}{(optional) the title of the plot.}
}
\value{
This function returns a bar plot that is ordered similarly across samples.
}
\description{
This function plots a bar plot from a \code{phyloseq} object.
The author of this fuction is GitHub user \href{https://github.com/pjames1}{pjames1} in \href{https://github.com/joey711/phyloseq/issues/442}{this thread}.
}
\examples{
plot_ordered_bar(ps)
plot_ordered_bar(ps, x = "Sample_ID", y = "Abundance", fill = Genus, title = "Sample Abundance")
}
