% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reltools.R
\docType{package}
\name{reltools}
\alias{reltools}
\alias{reltools-package}
\title{reltools: Microbiome Amplicon Analysis and Visualization.}
\description{
The \code{reltools} package provides functions for analysis and visualization of microbiome and proteome datasets.
It also includes functions for modeling ecological community assembly in datasets.
}
\section{Microbiome Functions}{

The microbiome functions are functions for modifying \code{phyloseq} objects.
}

\section{Proteome Functions}{

The proteome functions are functions for modifying \code{phyloseq} objects that contain count
or intensity data generated on a mass spec. In the future these functions will also operate on
\code{SummarizedExperiment} objects.
}

