% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_relman}
\alias{theme_relman}
\title{A custom theme for the Relman Lab}
\usage{
theme_relman()
}
\arguments{
\item{NULL}{}
}
\description{
This theme is inspired by the unique aesthetic sensibilities of David Relman, M.D.
}
\examples{

Make a Plot
p <- ggplot(data.frame(A =rnorm(100), B = rnorm(100)))
p <- p + geom_point(aes(x = A, y = B), color = "#FF56D6", size =3)
p <- p + ggtitle("A plot")
p <- p + theme_relman()
p

}
