% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{trial.mat}
\alias{trial.mat}
\title{Convert data.frame to matrix with Subjects on rows and items on columns}
\usage{
trial.mat(df, dv = NULL)
}
\arguments{
\item{df}{dataframe to convert to a matrix}

\item{dv}{name of column with observations to be put on columns}
}
\value{
matrix with Subjects on the rows and items on the column
}
\description{
Convert data.frame to matrix with Subjects on rows and items on columns
}
\examples{
trial.mat(data.frame(Subject=rep(1:2, each=3), 1:6))
trial.mat(data.frame(Subject=rep(1:2, each=3), a=1:6, b=11:16), dv='b')
}

